requirements.txt
asgiref==3.5.0
autopep8==2.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==6.3.2
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
packaging==21.3
psycopg2-binary==2.9.3
pycodestyle==2.10.0
PyJWT==1.7.1
pyparsing==3.0.7
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==6.0.0
zipp==3.7.0
